hybrid_agent_v2/requirements.txt
langgraph
chromadb
langchain
langchain-community
langchain-google-genai
google-generativeai
openai-whisper
tqdm
moviepy
python-dotenv

